Extract the numeric budget from the budget text string (if we can!)
The "budget" field in the marketplace data is a text string which can be (almost)
completely freeform. We want to be able to see the budget as a number, so we can
more easily see who is the biggest spender on the digital marketplace, or
order the marketplace data by budget, or any kind of number-based shenanigans.

Extracting the numerical budget is *hard* and a bit inexact, but
`Opportunity#extracted_budget` has a decent go.

It ignores day rates.
If a budget is expressed as a range, it uses the largest number.

To run this task, you will need to follow the setup instructions and run the
Rake task `add_opportunities:extract_budgets`. The budget string and the
extracted budget are in the final two columns.